ml-service/requirements.txt
pydantic>=2.0.0
httpx>=0.24.0
ollama>=0.4.0
bilibili-api-python>=16.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
